codigo correcto POR FIN CON SU PU

+++ src/main/kotlin/Articulos.kt
@@ -1,4 +1,4 @@
 import java.sql.Connection
 
-data class Articulos(val id_articulo: Int = -1, val nombre:String, val comentario:String, val precio:Double) {
+data class Articulos(val id_articulo: Int = -1, val nombre:String, val comentario:String, val precio:Int) {
 }
